actualizanco el tipo de dato de la llave primaria  para producto y fabricante en su clase repositorio

+++ src/main/java/com/isai/democomputerstore/app/repository/MakerRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface MakerRepository
-        extends JpaRepository<Maker, Long> {
+        extends JpaRepository<Maker, Integer> {
 }
